+++ embed_xray_other.go
@@ -1,0 +1,6 @@
+//go:build !windows
+
+package main
+
+var xrayEXE []byte
+var geoipDAT []byte