+++ embed_xray_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+
+package main
+
+import _ "embed"
+
+//go:embed assets/xray/xray.exe
+var xrayEXE []byte
+
+//go:embed assets/xray/geoip.dat
+var geoipDAT []byte